Ball counter Update Physics_ball.pde
Added a ball counter in the top left corner for when your trying to ruin your computer. Added a scale factor, when you turn down the size of the objects you can increase the scalefactor to make them move slower as if you were standing farther away. Added a new shade of purple.