The big update 4.pde
Added improved placement of particles added inputs for gui as well as removed some of the growing hotkeys. I added a orbital placement mode so when you have only one particle to be placed and orbital mode on. the program will change the velocity of the closest particle and place this one with enough velocity to orbit each oher.